[ads][CodeHealth] Convert Insert and BuildInsertSql to free functions in database tables
Clang's readability-convert-member-functions-to-static check flagged
Insert and BuildInsertSql across all 15 database table classes because
neither function accesses instance state. Moving them to the anonymous
namespace removes them from the class interface entirely. Where Insert
is called cross-file, it is instead exposed as a named free function in
the relevant util file. The four dead table member variables on Campaigns
that were replaced by these util free functions are also removed.

No behavioral changes.
