+++ components/brave_ads/core/internal/BUILD.gn
@@ -554,8 +554,12 @@ static_library("internal") {
     "creatives/creatives_info.h",
     "creatives/dayparts_database_table.cc",
     "creatives/dayparts_database_table.h",
+    "creatives/dayparts_database_table_util.cc",
+    "creatives/dayparts_database_table_util.h",
     "creatives/geo_targets_database_table.cc",
     "creatives/geo_targets_database_table.h",
+    "creatives/geo_targets_database_table_util.cc",
+    "creatives/geo_targets_database_table_util.h",
     "creatives/new_tab_page_ads/creative_new_tab_page_ad_info.cc",
     "creatives/new_tab_page_ads/creative_new_tab_page_ad_info.h",
     "creatives/new_tab_page_ads/creative_new_tab_page_ad_wallpaper_type.h",
@@ -584,6 +588,8 @@ static_library("internal") {
     "creatives/notification_ads/notification_ad_manager.h",
     "creatives/segments_database_table.cc",
     "creatives/segments_database_table.h",
+    "creatives/segments_database_table_util.cc",
+    "creatives/segments_database_table_util.h",
     "database/database.cc",
     "database/database.h",
     "database/database_maintenance.cc",
