+++ components/brave_ads/core/internal/account/deposits/deposits_database_table_util.h
@@ -6,12 +6,21 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_DEPOSITS_DEPOSITS_DATABASE_TABLE_UTIL_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_DEPOSITS_DEPOSITS_DATABASE_TABLE_UTIL_H_
 
+#include <map>
+#include <string>
+
 #include "brave/components/brave_ads/core/internal/account/deposits/deposit_info.h"
+#include "brave/components/brave_ads/core/internal/creatives/creative_deposit_info.h"
 #include "brave/components/brave_ads/core/mojom/brave_ads.mojom-forward.h"
 
 namespace brave_ads::database::table {
 
 DepositInfo DepositFromMojomRow(const mojom::DBRowInfoPtr& mojom_db_row);
+
+// Appends SQL actions to insert `deposits` into the `deposits` table.
+void InsertDeposits(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
+                    const std::map</*creative_instance_id*/ std::string,
+                                   CreativeDepositInfo>& deposits);
 
 }  // namespace brave_ads::database::table
