+++ components/brave_ads/core/internal/account/tokens/confirmation_tokens/confirmation_tokens_database_table.h
@@ -6,8 +6,6 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_TOKENS_CONFIRMATION_TOKENS_CONFIRMATION_TOKENS_DATABASE_TABLE_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_TOKENS_CONFIRMATION_TOKENS_CONFIRMATION_TOKENS_DATABASE_TABLE_H_
 
-#include <string>
-
 #include "base/functional/callback.h"
 #include "brave/components/brave_ads/core/internal/account/tokens/confirmation_tokens/confirmation_token_info.h"
 #include "brave/components/brave_ads/core/internal/database/database_table_interface.h"
@@ -28,9 +26,6 @@ class ConfirmationTokens final : public TableInterface {
   void Save(const ConfirmationTokenList& confirmation_tokens,
             ResultCallback callback);
 
-  void Insert(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
-              const ConfirmationTokenList& confirmation_tokens);
-
   void Delete(const ConfirmationTokenInfo& confirmation_token,
               ResultCallback callback);
   void DeleteAll(ResultCallback callback);
@@ -42,10 +37,6 @@ class ConfirmationTokens final : public TableInterface {
   void Migrate(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
                int to_version) override;
 
- private:
-  std::string BuildInsertSql(
-      const mojom::DBActionInfoPtr& mojom_db_action,
-      const ConfirmationTokenList& confirmation_tokens) const;
 };
 
 }  // namespace brave_ads::database::table
