+++ components/brave_ads/core/internal/account/tokens/payment_tokens/payment_tokens_database_table.h
@@ -6,8 +6,6 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_TOKENS_PAYMENT_TOKENS_PAYMENT_TOKENS_DATABASE_TABLE_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_TOKENS_PAYMENT_TOKENS_PAYMENT_TOKENS_DATABASE_TABLE_H_
 
-#include <string>
-
 #include "base/functional/callback.h"
 #include "brave/components/brave_ads/core/internal/account/tokens/payment_tokens/payment_token_info.h"
 #include "brave/components/brave_ads/core/internal/database/database_table_interface.h"
@@ -27,9 +25,6 @@ class PaymentTokens final : public TableInterface {
  public:
   void Save(const PaymentTokenList& payment_tokens, ResultCallback callback);
 
-  void Insert(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
-              const PaymentTokenList& payment_tokens);
-
   void Delete(const PaymentTokenInfo& payment_token, ResultCallback callback);
   void Delete(const PaymentTokenList& payment_tokens, ResultCallback callback);
   void DeleteAll(ResultCallback callback);
@@ -40,10 +35,6 @@ class PaymentTokens final : public TableInterface {
   void Create(const mojom::DBTransactionInfoPtr& mojom_db_transaction) override;
   void Migrate(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
                int to_version) override;
-
- private:
-  std::string BuildInsertSql(const mojom::DBActionInfoPtr& mojom_db_action,
-                             const PaymentTokenList& payment_tokens) const;
 };
 
 }  // namespace brave_ads::database::table
