+++ components/brave_ads/core/internal/account/transactions/transactions_database_table.h
@@ -6,8 +6,6 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_TRANSACTIONS_TRANSACTIONS_DATABASE_TABLE_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_ACCOUNT_TRANSACTIONS_TRANSACTIONS_DATABASE_TABLE_H_
 
-#include <string>
-
 #include "base/functional/callback.h"
 #include "brave/components/brave_ads/core/internal/account/tokens/payment_tokens/payment_token_info.h"
 #include "brave/components/brave_ads/core/internal/account/transactions/transaction_info.h"
@@ -42,12 +40,6 @@ class Transactions final : public TableInterface {
   void Migrate(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
                int to_version) override;
 
- private:
-  void Insert(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
-              const TransactionList& transactions);
-
-  std::string BuildInsertSql(const mojom::DBActionInfoPtr& mojom_db_action,
-                             const TransactionList& transactions) const;
 };
 
 }  // namespace brave_ads::database::table
