+++ components/brave_ads/core/internal/creatives/campaigns_database_table.h
@@ -6,22 +6,11 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_CAMPAIGNS_DATABASE_TABLE_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_CAMPAIGNS_DATABASE_TABLE_H_
 
-#include <map>
-#include <string>
-
-#include "brave/components/brave_ads/core/internal/account/deposits/deposits_database_table.h"
 #include "brave/components/brave_ads/core/internal/creatives/creative_ad_info.h"
-#include "brave/components/brave_ads/core/internal/creatives/dayparts_database_table.h"
-#include "brave/components/brave_ads/core/internal/creatives/geo_targets_database_table.h"
-#include "brave/components/brave_ads/core/internal/creatives/segments_database_table.h"
 #include "brave/components/brave_ads/core/internal/database/database_table_interface.h"
 #include "brave/components/brave_ads/core/mojom/brave_ads.mojom-forward.h"
 
-namespace brave_ads {
-
-struct CreativeCampaignInfo;
-
-namespace database::table {
+namespace brave_ads::database::table {
 
 class Campaigns final : public TableInterface {
  public:
@@ -43,20 +32,8 @@ class Campaigns final : public TableInterface {
  private:
   void MigrateToV48(const mojom::DBTransactionInfoPtr& mojom_db_transaction);
   void MigrateToV52(const mojom::DBTransactionInfoPtr& mojom_db_transaction);
-
-  std::string BuildInsertSql(
-      const mojom::DBActionInfoPtr& mojom_db_action,
-      const std::map</*campaign_id*/ std::string, CreativeCampaignInfo>&
-          campaigns) const;
-
-  GeoTargets geo_targets_database_table_;
-  Dayparts dayparts_database_table_;
-  Segments segments_database_table_;
-  Deposits deposits_database_table_;
 };
 
-}  // namespace database::table
-
-}  // namespace brave_ads
+}  // namespace brave_ads::database::table
 
 #endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_CAMPAIGNS_DATABASE_TABLE_H_
