+++ components/brave_ads/core/internal/creatives/creative_ads_database_table.h
@@ -22,9 +22,6 @@ using GetCreativeAdCallback =
 
 class CreativeAds final : public TableInterface {
  public:
-  void Insert(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
-              const CreativeAdList& creative_ads);
-
   void GetForCreativeInstanceId(const std::string& creative_instance_id,
                                 GetCreativeAdCallback callback) const;
 
@@ -36,9 +33,6 @@ class CreativeAds final : public TableInterface {
  private:
   void MigrateToV48(const mojom::DBTransactionInfoPtr& mojom_db_transaction);
   void MigrateToV54(const mojom::DBTransactionInfoPtr& mojom_db_transaction);
-
-  std::string BuildInsertSql(const mojom::DBActionInfoPtr& mojom_db_action,
-                             const CreativeAdList& creative_ads) const;
 };
 
 }  // namespace brave_ads::database::table
