+++ components/brave_ads/core/internal/creatives/dayparts_database_table.h
@@ -6,25 +6,13 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_DAYPARTS_DATABASE_TABLE_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_DAYPARTS_DATABASE_TABLE_H_
 
-#include <map>
-#include <string>
-
-#include "base/containers/flat_set.h"
 #include "brave/components/brave_ads/core/internal/database/database_table_interface.h"
 #include "brave/components/brave_ads/core/mojom/brave_ads.mojom-forward.h"
 
-namespace brave_ads {
-
-struct CreativeDaypartInfo;
-
-namespace database::table {
+namespace brave_ads::database::table {
 
 class Dayparts final : public TableInterface {
  public:
-  void Insert(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
-              const std::map</*campaign_id*/ std::string,
-                             base::flat_set<CreativeDaypartInfo>>& dayparts);
-
   // TableInterface:
   void Create(const mojom::DBTransactionInfoPtr& mojom_db_transaction) override;
   void Migrate(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
@@ -32,15 +20,8 @@ class Dayparts final : public TableInterface {
 
  private:
   void MigrateToV48(const mojom::DBTransactionInfoPtr& mojom_db_transaction);
-
-  std::string BuildInsertSql(
-      const mojom::DBActionInfoPtr& mojom_db_action,
-      const std::map</*campaign_id*/ std::string,
-                     base::flat_set<CreativeDaypartInfo>>& dayparts) const;
 };
 
-}  // namespace database::table
-
-}  // namespace brave_ads
+}  // namespace brave_ads::database::table
 
 #endif  // BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_DAYPARTS_DATABASE_TABLE_H_
