+++ components/brave_ads/core/internal/creatives/notification_ads/creative_notification_ads_database_table.h
@@ -6,12 +6,8 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_NOTIFICATION_ADS_CREATIVE_NOTIFICATION_ADS_DATABASE_TABLE_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_NOTIFICATION_ADS_CREATIVE_NOTIFICATION_ADS_DATABASE_TABLE_H_
 
-#include <cstddef>
-#include <string>
-
 #include "base/check_op.h"
 #include "brave/components/brave_ads/core/internal/creatives/campaigns_database_table.h"
-#include "brave/components/brave_ads/core/internal/creatives/creative_ads_database_table.h"
 #include "brave/components/brave_ads/core/internal/creatives/notification_ads/creative_notification_ad_info.h"
 #include "brave/components/brave_ads/core/internal/database/database_table_interface.h"
 #include "brave/components/brave_ads/core/internal/segments/segment_types.h"
@@ -56,17 +52,9 @@ class CreativeNotificationAds final : public TableInterface {
  private:
   void MigrateToV48(const mojom::DBTransactionInfoPtr& mojom_db_transaction);
 
-  void Insert(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
-              const CreativeNotificationAdList& creative_ads);
-
-  std::string BuildInsertSql(
-      const mojom::DBActionInfoPtr& mojom_db_action,
-      const CreativeNotificationAdList& creative_ads) const;
-
   int batch_size_;
 
   Campaigns campaigns_database_table_;
-  CreativeAds creative_ads_database_table_;
 };
 
 }  // namespace brave_ads::database::table
