+++ components/brave_ads/core/internal/creatives/segments_database_table.h
@@ -6,10 +6,6 @@
 #ifndef BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_SEGMENTS_DATABASE_TABLE_H_
 #define BRAVE_COMPONENTS_BRAVE_ADS_CORE_INTERNAL_CREATIVES_SEGMENTS_DATABASE_TABLE_H_
 
-#include <map>
-#include <string>
-
-#include "base/containers/flat_set.h"
 #include "brave/components/brave_ads/core/internal/database/database_table_interface.h"
 #include "brave/components/brave_ads/core/mojom/brave_ads.mojom-forward.h"
 
@@ -17,10 +13,6 @@ namespace brave_ads::database::table {
 
 class Segments final : public TableInterface {
  public:
-  void Insert(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
-              const std::map</*creative_set_id*/ std::string,
-                             base::flat_set<std::string>>& segments);
-
   // TableInterface:
   void Create(const mojom::DBTransactionInfoPtr& mojom_db_transaction) override;
   void Migrate(const mojom::DBTransactionInfoPtr& mojom_db_transaction,
@@ -28,11 +20,6 @@ class Segments final : public TableInterface {
 
  private:
   void MigrateToV48(const mojom::DBTransactionInfoPtr& mojom_db_transaction);
-
-  std::string BuildInsertSql(
-      const mojom::DBActionInfoPtr& mojom_db_action,
-      const std::map</*creative_set_id*/ std::string,
-                     base::flat_set<std::string>>& segments) const;
 };
 
 }  // namespace brave_ads::database::table
